+++ paddle/fluid/prim/api/auto_code_generated/tensor_operants_gen.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,7 +15,9 @@
 import argparse
 
 import yaml
-from prim_base import EagerPrimAPI
+from eager_gen import BaseAPI
+
+indent = "  "
 
 eager_header_include = """// Generated by paddle/fluid/prim/api/auto_code_generated/tensor_operants_gen.py
 
@@ -139,6 +141,102 @@ static_source_end = """
 """
 
 
+class PrimTensorAPI(BaseAPI):
+    def __init__(self, api_item_yaml, prims=tuple()):
+        super().__init__(api_item_yaml, prims)
+
+    def get_api_func_name(self):
+        return self.api
+
+    # def is_inplace(self):
+    #     if self.inplace_map
+    #         return True
+    #     return False
+
+    def gene_tensor_operants_declaration(self):
+        api_func_name = self.get_api_func_name()
+
+        if api_func_name[-1] != '_':
+            return f"""{indent}{self.get_return_type()} {api_func_name}({self.get_declare_args()});\n
+"""
+        else:
+            return f"""{indent}{self.get_return_type(inplace_flag=True)} {api_func_name}({self.get_declare_args(inplace_flag=True)});\n
+"""
+
+    def get_func_input_args(self, inplace_flag=False):
+        input_args = []
+        for name in self.inputs['names']:
+            name = name.split('@')[0]
+            if inplace_flag and name in self.inplace_map.values():
+                input_args.append(name)
+            else:
+                input_args.append(name)
+        return input_args
+
+    def get_func_args(self, inplace_flag=False):
+        ad_func_args = self.get_func_input_args(inplace_flag)
+        for name in self.attrs['names']:
+            default_value = ''
+            if self.attrs['attr_info'][name][1] is not None:
+                default_value = ' = ' + self.attrs['attr_info'][name][1]
+            ad_func_args.append(name)
+
+        ad_func_args_str = ", ".join(ad_func_args)
+        return ad_func_args_str
+
+    def gene_eager_tensor_func_call(self):
+        api_func_name = self.get_api_func_name()
+
+        dygraph_ad_func_name = '::' + api_func_name + '_ad_func'
+        dygraph_ad_func_parameters = self.get_func_args()
+
+        return (
+            f"""return {dygraph_ad_func_name}({dygraph_ad_func_parameters});"""
+        )
+
+    def gene_eager_tensor_operants_implementation(self):
+        api_func_name = self.get_api_func_name()
+        # func decalaration
+        if api_func_name[-1] != '_':
+            api_code = f"""{self.get_return_type()} EagerTensorOperants::{api_func_name}({self.get_declare_args_nodefault()}) {{"""
+        else:
+            api_code = f"""{self.get_return_type(inplace_flag=True)} EagerTensorOperants::{api_func_name}({self.get_declare_args_nodefault(inplace_flag=True)}) {{"""
+
+        # func code
+        api_code += f"""
+{indent}{self.gene_eager_tensor_func_call()}\n}}\n
+"""
+        return api_code
+
+    def gene_static_tensor_func_call(self):
+        api_func_name = self.get_api_func_name()
+
+        prim_static_func_name = (
+            'paddle::prim::' + api_func_name + '<DescTensor>'
+        )
+        prim_static_func_parameters = self.get_func_args()
+
+        return f"""return {prim_static_func_name}({prim_static_func_parameters});"""
+
+    def gene_static_tensor_operants_implementation(self):
+        api_code = ""
+        indent = "  "
+        api_func_name = self.get_api_func_name()
+        # func decalaration
+        if api_func_name[-1] != '_':
+            api_code = f"""{self.get_return_type()} StaticTensorOperants::{api_func_name}({self.get_declare_args_nodefault()}) {{"""
+        else:
+            api_code = f"""{self.get_return_type(inplace_flag=True)} StaticTensorOperants::{api_func_name}({self.get_declare_args_nodefault(inplace_flag=True)}) {{"""
+
+        function_call = self.gene_static_tensor_func_call()
+        # func code
+        api_code += f"""
+{indent}{function_call}\n}}\n
+"""
+
+        return api_code
+
+
 def generate_tensor_operants_api(
     api_yaml_path,
     eager_header_path,
@@ -175,7 +273,7 @@ def generate_tensor_operants_api(
         api_prims = ('add', 'subtract', 'multiply', 'divide')
 
     for api in apis:
-        eager_api = EagerPrimAPI(api, api_prims)
+        eager_api = PrimTensorAPI(api, api_prims)
         if eager_api.is_prim_api:
             eager_header_file.write(
                 eager_api.gene_tensor_operants_declaration()
